Fixat dbcontext

+++ AllwinAPI/AllwinAPI/Model/Interfaces/IStop.cs
@@ -4,8 +4,8 @@
     {
         int StopId { get; set; }
         int RouteId { get; set; }
-        double Latitude { get; set; }
-        double Longitude { get; set; }
+        float Latitude { get; set; }
+        float Longitude { get; set; }
         string Name { get; set; }
         string Adress { get; set; }
     }
